+++ src/main/java/net/lordimass/dialogue/parameter/ParameterContextContributor.java
@@ -1,6 +1,0 @@
-package net.lordimass.dialogue.parameter;
-
-@FunctionalInterface
-public interface ParameterContextContributor {
-    void contribute(ParameterContext ctx);
-}