+++ src/main/java/net/lordimass/dialogue/parameter/ParameterProcessor.java
@@ -1,10 +1,9 @@
 package net.lordimass.dialogue.parameter;
 
-public final class ParameterProcessor<C> {
-
-    private final String key;
-    private final Class<C> contextType;
-    private final ParameterResolver<C> resolver;
+public abstract class ParameterProcessor<C> {
+    protected final String key;
+    protected final Class<C> contextType;
+    protected final ParameterResolver<C> resolver;
 
     public ParameterProcessor(String key, Class<C> contextType, ParameterResolver<C> resolver) {
         this.key = key;
@@ -12,15 +11,6 @@ public final class ParameterProcessor<C> {
         this.resolver = resolver;
     }
 
-    public String key() {
-        return key;
-    }
-
-    public boolean supports(ParameterContext ctx) {
-        return ctx.has(contextType);
-    }
-
-    public String resolve(ParameterContext ctx) {
-        return resolver.resolve(ctx.get(contextType));
-    }
+    abstract public boolean supports(ParameterContext ctx);
+    abstract public String resolve(ParameterContext ctx);
 }