+++ src/main/java/net/lordimass/dialogue/parameter/ParameterResolver.java
@@ -1,6 +1,4 @@
 package net.lordimass.dialogue.parameter;
 
-@FunctionalInterface
 public interface ParameterResolver<C> {
-    String resolve(C context);
 }